components admin y public terminados, error no trae datos

+++ src/app/admin/hotel/pages/list-hotel/list-hotel.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component, inject, signal } from '@angular/cor
 import { CommonModule } from '@angular/common';
 import { Hotel } from '../../../interfaces/hotel.interface';
 import { HotelService } from '../../../services/hotel.service';
-import { RouterLink } from "@angular/router";
+import { RouterLink } from '@angular/router';
 
 @Component({
   standalone: true,
@@ -12,7 +12,6 @@ import { RouterLink } from "@angular/router";
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class ListHotelPageComponent {
-
   private hotelService = inject(HotelService);
 
   hoteles = signal<Hotel[]>([]);
@@ -24,17 +23,16 @@ export class ListHotelPageComponent {
   loadAllHoteles() {
     this.hotelService.getAllHoteles().subscribe({
       next: (data) => this.hoteles.set(data),
-      error: (err) => console.error('Error cargando hoteles', err)
+      error: (err) => console.error('Error cargando hoteles', err),
     });
   }
 
   eliminarHotelPorId(id: number) {
-    this.hotelService.deleteById(id).subscribe({
-      next: () => {
-        alert('Seguro que quieres eliminar el hotel?')
-        this.loadAllHoteles()
-      },
-      error: err => console.log(err)
-    })
+    if (confirm('¿Seguro que quieres eliminar el hotel?')) {
+      this.hotelService.deleteById(id).subscribe({
+        next: () => this.loadAllHoteles(),
+        error: (err) => console.log(err),
+      });
+    }
   }
 }
